requirements.txt
dspy~=3.0.3
wandb~=0.19.8
numpy~=1.26.4
dataclasses~=0.6
datasets~=3.2.0
stemmer~=0.0.4
bm25s~=0.2.14
diskcache~=5.6.3
nltk~=3.9.1
spacy~=3.7.4
emoji~=2.15.0
syllapy~=0.7.2
immutabledict~=4.2.1
langdetect~=1.0.9
packaging~=24.2
sympy~=1.14.0
lark~=1.2.2
python-dotenv~=1.0.1